pyproject.toml
[project]
name = "umamusume-web-crawler"
version = "0.1.0"
description = "UmaMusume web crawler utilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.7.8",
    "google-api-python-client>=2.127.0",
    "httplib2==0.22.0",
    "langchain-mcp-adapters>=0.2.1",
    "markitdown[all]>=0.1.4",
    "mcp>=1.25.0",
    "mcp-server>=0.1.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.1",
    "starlette>=0.37.2",
    "uvicorn>=0.30.0",
]
